requirements/base.txt
django==3.2.*
daphne
djangorestframework
django-admin-honeypot
django-ipware
tomli